Distributed NP with Celery

+++ model/NepacProcess.py
@@ -128,7 +128,10 @@ class NepacProcess(object):
             #   Mission1-pVal1, Mission1-pVal2, Mission2-pVal2]
             # -------------------------------------------------------------------------
 
-            ex = self._processTimeDate(timeDate, timeDateToLocChl[timeDate])
+            ex = self._processTimeDate(timeDate,
+                                       timeDateToLocChl[timeDate],
+                                       self._missions,
+                                       self._outputDir)
             rowsPerTimeDate = []
 
             for i, (missionKey, missionVals) in enumerate(ex.items()):
@@ -137,7 +140,8 @@ class NepacProcess(object):
                     # First time seeing these keys, new row.
                     if i == 0:
                         newRow = []
-                        newRow.extend(list(rowKey))
+                        rowKeyTuple = tuple(rowKey.split(","))
+                        newRow.extend(list(rowKeyTuple))
                         newRow.extend(rowValues)
                         rowsPerTimeDate.append(newRow)
 
@@ -254,18 +258,22 @@ class NepacProcess(object):
     #       }
     # }
     # -------------------------------------------------------------------------
-    def _processTimeDate(self, timeDate, locsChls):
+    @staticmethod
+    def _processTimeDate(timeDate, locsChls, missions, outputDir):
 
         print('Processing', timeDate)
 
         valuesPerMissionDict = {}
 
-        for mission in self._missions:
+        for mission in missions:
 
-            valuesPerMissionDict[mission] = \
-                self._processMission(mission,
-                                     timeDate,
-                                     locsChls)
+            dictOutput = \
+                NepacProcess._processMission(mission,
+                                             timeDate,
+                                             locsChls,
+                                             missions,
+                                             outputDir)
+            valuesPerMissionDict.update(dictOutput)
 
         sortedValuesPerMissionDict = dict(
             sorted(valuesPerMissionDict.items(),
@@ -278,7 +286,8 @@ class NepacProcess(object):
     #
     # This can be distributed.
     # -------------------------------------------------------------------------
-    def _processMission(self, mission, timeDate, locsChls):
+    @staticmethod
+    def _processMission(mission, timeDate, locsChls, missions, outputDir):
 
         print('Processing mission', mission)
         print('timeDate: ', timeDate)
@@ -291,7 +300,7 @@ class NepacProcess(object):
         year = int(timeDate[1][4:])
         dt = datetime.datetime(year, month, day, hour, minute)
 
-        ocr = OceanColorRetriever(mission, dt, self._outputDir)
+        ocr = OceanColorRetriever(mission, dt, outputDir)
         missionFile = ocr.run()
 
         # Instantiate a GeospatialImageFile to access its data sets.
@@ -305,7 +314,7 @@ class NepacProcess(object):
         # probably not be distributed.  There would be multiple processes
         # trying to open the same image.
         # ---
-        dataSets = self._missions[mission]
+        dataSets = missions[mission]
         nepacOutputDict = {}
 
         for sub in sorted(subs):
@@ -340,11 +349,13 @@ class NepacProcess(object):
                     # Byte form to float
                     [val] = struct.unpack('f', val)
 
-                    timeDateLocChlKey = (timeDate[0],
-                                         timeDate[1],
-                                         locChl[0],
-                                         locChl[1],
-                                         locChl[2])
+                    timeDateLocChlKey = (
+                        timeDate[0],
+                        timeDate[1],
+                        locChl[0],
+                        locChl[1],
+                        locChl[2])
+                    timeDateLocChlKey = ','.join(timeDateLocChlKey)
 
                     # New key to be made, appends vals
                     if timeDateLocChlKey not in nepacOutputDict:
@@ -359,6 +370,10 @@ class NepacProcess(object):
                 dsImage = None
                 os.remove(name)
 
-        os.remove(missionFile)
-        return nepacOutputDict
+        if os.path.exists(missionFile):
+            os.remove(missionFile)
 
+        nepacMissionOutput = {}
+        nepacMissionOutput[mission] = nepacOutputDict
+
+        return nepacMissionOutput
